pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "srg-rm-copilot"
version = "0.1.0"
description = "Production-ready Python package + CLI + ETL for Wheelhouse data with AI-powered development automation"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "SRG Team"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "pyarrow>=15.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "tenacity>=8.0.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-httpx>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-httpx>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "black>=24.0.0",
    "pre-commit>=3.0.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/srg/srg-rm-copilot"
Repository = "https://github.com/srg/srg-rm-copilot.git"
Issues = "https://github.com/srg/srg-rm-copilot/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/srg_rm_copilot"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.per-file-ignores]
"tests/*" = ["E501"]

[tool.black]
target-version = ['py312']
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/srg_rm_copilot",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
